Change kubeconfig testing name and move into a separate file
