+++ tests/units/providers/test_openshift_provider.py
@@ -225,168 +225,3 @@ class TestOpenshiftProviderProcessArtifactData(OpenshiftProviderTestMixin, unitt
             op._process_artifact_data, 'foo', artifact_data)
 
 
-class TestOpenshiftProviderParseKubeconfData(OpenshiftProviderTestMixin, unittest.TestCase):
-
-    def test_parse_kubeconf_data_insecure(self):
-        """
-        Test parsing kubeconf data with current context containing
-        cluster, user, namespace info and skipping tls verification
-        """
-        kubecfg_data = {
-            'current-context': 'context2',
-            'contexts': [
-                {
-                    'name': 'context1',
-                },
-                {
-                    'name': 'context2',
-                    'context': {
-                        'cluster': 'cluster1',
-                        'user': 'user1',
-                        'namespace': 'namespace1'
-                    }
-                }
-            ],
-            'clusters': [
-                {
-                    'name': 'cluster1',
-                    'cluster': {
-                        'insecure-skip-tls-verify': 'true',
-                        'server': 'server1'
-                    }
-                }
-            ],
-            'users': [
-                {
-                    'name': 'user1',
-                    'user': {
-                        'token': 'token1'
-                    }
-                }
-            ]
-        }
-
-        op = self.get_oc_provider()
-        self.assertEqual(op._parse_kubeconf_data(kubecfg_data),
-                         {'providerapi': 'server1',
-                          'accesstoken': 'token1',
-                          'namespace': 'namespace1',
-                          'providertlsverify': False,
-                          'providercafile': None})
-
-    def test_parse_kubeconf_data_cafile(self):
-        """
-        Test parsing kubeconf data with current context containing
-        cluster, user, namespace info and certificate-authority
-        """
-        kubecfg_data = {
-            'current-context': 'context2',
-            'contexts': [
-                {
-                    'name': 'context1',
-                },
-                {
-                    'name': 'context2',
-                    'context': {
-                        'cluster': 'cluster1',
-                        'user': 'user1',
-                        'namespace': 'namespace1'
-                    }
-                }
-            ],
-            'clusters': [
-                {
-                    'name': 'cluster1',
-                    'cluster': {
-                        'certificate-authority': '/foo/bar',
-                        'server': 'server1'
-                    }
-                }
-            ],
-            'users': [
-                {
-                    'name': 'user1',
-                    'user': {
-                        'token': 'token1'
-                    }
-                }
-            ]
-        }
-
-        op = self.get_oc_provider()
-        self.assertEqual(op._parse_kubeconf_data(kubecfg_data),
-                         {'providerapi': 'server1',
-                          'accesstoken': 'token1',
-                          'namespace': 'namespace1',
-                          'providertlsverify': True,
-                          'providercafile': '/foo/bar'})
-
-    def test_parse_kubeconf_data_no_context(self):
-        """
-        Test parsing kubeconf data with missing context data for
-        current context.
-        """
-        kubecfg_data = {
-            'current-context': 'context2',
-            'contexts': [
-                {
-                    'name': 'context1',
-                }
-            ],
-            'clusters': [
-                {
-                    'name': 'cluster1',
-                    'cluster': {
-                        'server': 'server1'
-                    }
-                }
-            ],
-            'users': [
-                {
-                    'name': 'user1',
-                    'user': {
-                        'token': 'token1'
-                    }
-                }
-            ]
-        }
-
-        op = self.get_oc_provider()
-        self.assertRaises(ProviderFailedException,
-                          op._parse_kubeconf_data, kubecfg_data)
-
-    def test_parse_kubeconf_data_no_user(self):
-        """
-        Test parsing kubeconf data with missing user data in current
-        context.
-        """
-        kubecfg_data = {
-            'current-context': 'context2',
-            'contexts': [
-                {
-                    'name': 'context1',
-                },
-                {
-                    'name': 'context2',
-                    'context': {
-                        'cluster': 'cluster1',
-                        'user': 'user1',
-                        'namespace': 'namespace1'
-                    }
-                }
-            ],
-            'clusters': [
-                {
-                    'name': 'cluster1',
-                    'cluster': {
-                        'server': 'server1'
-                    }
-                }
-            ],
-            'users': [
-            ]
-        }
-
-        op = self.get_oc_provider()
-        self.assertRaises(ProviderFailedException,
-                          op._parse_kubeconf_data, kubecfg_data)
